pyproject.toml
[build-system]
requires = [ "poetry>=0.12",]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "liquidpy"
version = "0.6.0"
description = "A port of liquid template engine for python"
authors = [ "pwwang <pwwang@pwwang.com>",]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/pwwang/liquidpy"
repository = "https://github.com/pwwang/liquidpy"
include = ["liquid/tags/grammar.lark", "liquid/python/tags/grammar.lark"]

[[tool.poetry.packages]]
include = "liquid"

[tool.poetry.dependencies]
python = "^3.6"
diot = "*"
lark-parser = "^0.9"
rich = "6.*"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
python-dateutil = "^2"
